B.0.8.0B - Begin of Radio Button Work
Beginning of development involving radio buttons for variable max power thresholds (for Winch, Crank, and Drive Train).

In the middle of writing a command based version of this after Adam began an iterative one that did not end up working. Based off of the autonomous radio buttons found in the 2018 code base.

Co-Authored-By: bnjimenez <bnjimenez@users.noreply.github.com>