B0.10.2B - Reflective Tape Tracking
Created a new limelight pipeline (#5) titled Reflective_Tape_Hatches to allow for the tracking and scoring of hatches.
Accomplished a drive-in routine using the "Back" button on the operator controller with the previously created LineTrackCommand.
Tuned the PIDHatchMechanism so that it will rotate less, this was done so that the hatch does not naturally want to fall off of the tongue.