B0.10.3S - Auto-Shifting System Logic Established
The if statements to determine if the robot should shift into high gear or stay in low gear have been implemented and tested, however it DOES NOT control the actual shifting of the robot's transmissions. This is to be added when the design for the shifting system has been finalized.