B0.12.4B - Gear Shifter
Created code to manually control the gear shifter on the robot. The gear shifter is identified by Talon 12 within the DRIVE_SUBSYSTEM and is controlled using PID within the new default command for DRIVE_SUBSYSTEM titled: "ShiftDriveCommand". This allows the driver to swap from the home position of Low Gear to High Gear while holding Button #9 on the left Joystick.