B0.13.1B - Subsystem Check
Tested most subsystems including the non-finalized versions of:
The climb, the drive train, the dart, and the tongue.

Placed the control of the dart onto the operator controller. The Y Button deploys the dart from home and the A Button currently brings the dart back home.

Had to reverse the drive direction of the tongue.

Made it so that to climb with the current setup, you must be holding both the Left and Right Bumpers on the operator controller.