B1.0.5B - End of Chestnut Hill (Day 3)
Added left and right autonomous routines to score in the first cargo ship bay.
Slowed down the dart motor after a mechanical repair was made which resulted in an increase in speed.

All code is based off of what was ran in quarterfinals match two with a couple of adjustments made to the autonomous in terms of rotation and drive distance for Left/Right Face Cargo Ship Level Two. In addition to the adjustments, a step has been added which backs up 8 inches to provide the vision tracking with more room for adjustment.