S1.1.1B - NEO Motor Climb Test
Tested the competition climb code with a substitution of the NEO motors and Spark Max speed controllers instead of CIMS and Talon SRX speed controllers.

Also added a changeClimbBrakeMode method to be called in disabledInit to place the climb into coast mode as well as autonomousInit and teleopInit to place the climb into brake mode.