S1.1.2B - Two NEO Motor Climb Test
Successfully climbed with two NEO motors.
The climb is currently capable of driving in both directions based off of the left joystick y-axis on the operator controller. This climb was performed using two NEO motors with a full range of power (-1.0 to 1.0) and the battery mounted at the front of the robot to prevent the robot from tipping backwards.