S1.1.4B - SmartDashboard Feedback
Added a boolean box to the smart dashboard which is green while the robot is being commanded autonomously, then turns to red as soon as the drivers regain control.
From then on, the boolean box will turn green when the tongue is deployed signaling to the driver that they are ready to back out.
Added a changeDriveBrakeMode to keep the drive train in brake mode throughout autonomous and teleop, but not when it is disabled.