B1.1.4B - Integrating Sams-Environment
Added in the NEO/SparkMax control based off of Storm Robotics' code as well as the SmartDashboard feedback code that was developed on Sams-Environment.
Also included is the changeDriveBrakeMode and changeClimbBrakeMode code to automatically swap when going from disabledInit(), autonomousInit(), and teleopInit().
Slimmed down the amount of data being streamed through updateOI to the SmartDashboard.