B1.2.1B - End of Montgomery Day 2
Adjustments made at the beginning of the day that were carried through up until the end of the day (up to our 7th match to end the day).
Lowered the drive distance for both the left and right face cargo ship.
Tuned the LineTrackCommand using the graphing calculator tool: desmos (https://www.desmos.com/calculator).
Made changes to the resolution and FPS of the Microsoft LifeCam to stay within the 4Mb/s bandwidth limit.